Questions entered on frontend now sent to database
