+++ .vscode/launch.json
@@ -1,0 +1,14 @@
+{
+    "configurations": [
+        {
+            "type": "java",
+            "name": "Spring Boot-BackEndApplication<back-end>",
+            "request": "launch",
+            "cwd": "${workspaceFolder}",
+            "console": "internalConsole",
+            "mainClass": "com.crypto_tutor.BackEndApplication",
+            "projectName": "back-end",
+            "args": ""
+        }
+    ]
+}
